added old m7 algo and its logic from old fork
* remaining issue: the program may tend to crash on start with m7. It
may-be required to restart ccminer several time before it start mining.
(This is under investigation)

+++ cuda_helper.h
@@ -24,7 +24,7 @@
 #endif
 
 extern "C" short device_map[MAX_GPUS];
-extern "C"  long device_sm[MAX_GPUS];
+extern "C"  uint32_t device_sm[MAX_GPUS];
 
 extern int cuda_arch[MAX_GPUS];
